pyproject.toml
[project]
name = "automation"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.14.2",
    "celery>=5.5.3",
    "crispy-bootstrap5>=2025.6",
    "django>=5.2.7",
    "django-anymail>=13.1",
    "django-autocomplete-light>=3.12.1",
    "django-celery-results>=2.6.0",
    "django-ckeditor>=6.7.3",
    "django-crispy-forms>=2.4",
    "pillow>=12.0.0",
    "python-decouple>=3.8",
    "redis>=7.0.0",
    "requests>=2.32.5",
    "yfinance>=0.2.66",
]
